requirements.txt
accelerate==1.10.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.4.0
alembic==1.13.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
bcrypt==4.3.0
bitsandbytes==0.46.1
black==23.11.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
coverage==7.10.1
croniter==6.0.0
cryptography==45.0.6
git+https://github.com/huggingface/diffusers.git@03c3f69aa57a6cc2c995d41ea484195d719a240a
distlib==0.4.0
ecdsa==0.19.1
fastapi==0.104.1
ffmpy==0.6.1
filelock==3.18.0
flake8==7.3.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
GPUtil==1.4.0
gradio==5.38.2
gradio_client==1.11.0
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.34.3
identify==2.6.13
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==5.12.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
mypy==1.7.1
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.2
nvidia-ml-py==12.575.51
opencv-python==4.12.0.88
orjson==3.11.1
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
propcache==0.3.2
psutil==5.9.6
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pynvml==12.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.18
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
rsa==4.9.1
ruff==0.12.5
safehttpx==0.1.6
safetensors==0.6.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.13.3
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.55.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
virtualenv==20.34.0
# wan_cli package is installed locally, not from git
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==12.0
yarl==1.20.1
zipp==3.23.0
mkdocs>=1.5.0
mkdocs-material>=9.0.0
mkdocs-mermaid2-plugin>=0.6.0
beautifulsoup4>=4.12.0
jsonschema>=4.17.0
cerberus>=1.3.4
# Model Orchestrator dependencies
boto3>=1.34.0
botocore>=1.34.0
keyring>=24.0.0
tomli>=2.0.0
tomli-w>=1.0.0